+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add `excludeFiles` option
 - Add `excludeFolders` option
+- Add `includeDotFiles` option
 
 ## 1.6.5 (2023-05-27)